pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vibe-piper"
version = "0.1.0"
description = "Declarative data pipeline, integration, quality, transformation, and activation library"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Vibe Piper Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "httpx>=0.27.0",
    "jinja2>=3.1.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "pydantic>=2.10.0",
    "pyarrow>=23.0.0",
    "python-snappy>=0.7.3",
    "typer>=0.12.0",
    "rich>=13.7.0",
]

[project.scripts]
vibepiper = "vibe_piper.cli.main:cli"

[project.optional-dependencies]
files = [
    "pandas>=2.0.0",
    "pyarrow>=12.0.0",
    "openpyxl>=3.0.0",
    "python-snappy>=0.6.0",
]
all = [
    "vibe-piper[files,postgres,mysql,snowflake,bigquery]",
]
postgres = [
    "psycopg2-binary>=2.9.0",
]
mysql = [
    "mysql-connector-python>=8.0.0",
]
snowflake = [
    "snowflake-connector-python>=3.0.0",
]
bigquery = [
    "google-cloud-bigquery>=3.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unreachable = true
strict_equality = true
pretty = true
show_error_context = true
show_column_numbers = true
show_error_codes = true
exclude = [
    "examples/",
]

[[tool.mypy.overrides]]
module = "tests.*"
strict = false
check_untyped_defs = false

[[tool.mypy.overrides]]
module = "examples.*"
strict = false
check_untyped_defs = false

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
]
ignore = [
    "E501",  # line length is handled by the formatter
    "E402",  # allow non-top-level imports in bootstrap-style modules/tests
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ARG001", "E402", "E731", "F401", "F821", "F841"]
"src/vibe_piper/__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "integration: marks tests as integration tests (requires database)",
]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]

[dependency-groups]
dev = [
    "mypy>=1.10.0",
    "pre-commit>=4.0.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
]

docs = [
    "myst-parser>=2.0.0",
    "sphinx>=7.4.0",
    "sphinx-autobuild>=2024.10.3",
    "sphinx-autodoc-typehints>=2.3.0",
    "sphinx-rtd-theme>=2.0.0",
]
